Added Fullscreen, Background Color & Scroll support
The window is now capable of going in & out of fullscreen, with the usage
of the F11 key. Along with that, additional support for resizing was
added.

Support from OpenGL was added to make the window rendable. Also, color
was added to the background of the window.

Basic support for scroll was added into the engine.io.Input class.

This project is based on the Youtuber CodingAP series about LWJGL 3

The link to the fourth video:
https://www.youtube.com/watch?v=plsmoqJ64ho