Added basic Rendering
Added package engine.math and in it the class Vector3f, which represents
a vector of three floats.

Added package engine.graphics and its Classes Mesh, Renderer & Vertex.
Class Vertex represents a Vector3f for now.
Class Mesh presents a mesh/object. Its points coordinates and the indeces
for drawing were added.
Clash Renderer is used to render a specified mesh object. The drawing
shapes are specified as triangles.

In main a simple rectangle mesh is defined and the renderer is called to
draw it.

This project is based on the Youtuber CodingAP series about LWJGL 3

The link to the fifth video:
https://www.youtube.com/watch?v=1F9shq6KubY