Added shaders
Added package engine.utils and in it the class FileUtils, which is used
to read the content of the shader file and conver it into a String.
The input method is different for the one shown in the video tutorial.

Added the resources folder with the two shader files: mainVertex & 
mainFragment.

In Main, the shader was called and added to the simple rectangle.

This project is based on the Youtuber CodingAP series about LWJGL 3

The link to the sixth video:
https://www.youtube.com/watch?v=Ssxd17AOgzM