Enable coloring from the mesh and not by the shaderes
Added a VBO for colors, named cbo. In cbo the colors of the mesh are
stored. These colors are passed from the constructor.

Changed the mainVertex shader in order to take the color from the mesh,
and not by its position.

This project is based on the Youtuber CodingAP series about LWJGL 3

The link to the seventh video:
https://www.youtube.com/watch?v=90sv3ZdTXgw